Minified the JS with Terser CLI
